Address review comments
- Add explicit InsecureSkipVerify field to GRPCConfig and
  FlowAggregatorConfig. Instead of implicitly skipping TLS verification
  when no CA cert is provided, the field must now be set explicitly.
  A warning is logged during initialization when it is enabled.

- Use string names for the flowTypes URL query parameter (intra-node,
  inter-node, to-external, from-external) instead of integer values.
  This matches the direction parameter convention and makes URLs
  human-readable.

- Inline the useFlowStream hook into flowvisibility.tsx (its only
  consumer) and remove the separate src/hooks/use-flow-stream.ts file.
